Adds Bluetooth Interface
Adds an bluetooth interface that creates BLE service with a write
characteristic. The service advertises itself so that clients can
connect via Bluetooth Low Energy.

The characteristic is defined as a write characteristic to receive data
from clients. Notification about newly received packages together with
the parsed data is provided via a callback